+++ Sources/ReadAlign/Rules.swift
@@ -29,7 +29,6 @@ public struct Rules: Codable, Sendable {
     public let pieceSeconds: TimeInterval
     public let shortestPieceShare: Double
     public let pauseSeconds: TimeInterval
-    public let sameMoment: TimeInterval
     public let askAgainTrims: [TimeInterval]
     public let shortestWorthAskingAgain: TimeInterval
 
@@ -59,7 +58,6 @@ public struct Rules: Codable, Sendable {
         case pieceSeconds = "piece_seconds"
         case shortestPieceShare = "shortest_piece_share"
         case pauseSeconds = "pause_seconds"
-        case sameMoment = "same_moment"
         case askAgainTrims = "ask_again_trims"
         case shortestWorthAskingAgain = "shortest_worth_asking_again"
     }